chore: refactor to drawer style
